Update functest requirements because of Django 2.0 support
