+++ tests/core/test_accounts.py
@@ -1,9 +1,15 @@
 from django.conf import settings as django_settings
+from django.contrib import auth
 from django.contrib.auth import authenticate
+from django.contrib.auth.models import AnonymousUser
 from django.shortcuts import resolve_url
+from wiki.conf import settings as wiki_settings
+from wiki.models import reverse
 
-from ..base import (ArticleWebTestUtils, DjangoClientTestBase,
-                    RequireRootArticleMixin)
+from ..base import (SUPERUSER1_PASSWORD, SUPERUSER1_USERNAME,
+                    ArticleWebTestUtils, DjangoClientTestBase,
+                    RequireRootArticleMixin, TestBase, wiki_override_settings)
+from ..testdata.models import CustomUser
 
 
 class AccountUpdateTest(RequireRootArticleMixin, ArticleWebTestUtils, DjangoClientTestBase):
@@ -57,3 +63,63 @@ class AccountUpdateTest(RequireRootArticleMixin, ArticleWebTestUtils, DjangoClie
                 password=example_data['password1']
             )
         )
+
+
+class UpdateProfileViewTest(RequireRootArticleMixin, ArticleWebTestUtils, DjangoClientTestBase):
+
+    def test_update_profile(self):
+        self.client.post(
+            resolve_url('wiki:profile_update'),
+            {"email": "test@test.com", "password1": "newPass", "password2": "newPass"},
+            follow=True
+        )
+
+        test_auth = authenticate(username='admin', password='newPass')
+
+        self.assertNotEqual(test_auth, None)
+        self.assertEqual(test_auth.email, 'test@test.com')
+
+
+@wiki_override_settings(ACCOUNT_HANDLING=True)
+class LogoutViewTests(RequireRootArticleMixin, DjangoClientTestBase):
+
+    def test_logout_account_handling(self):
+        self.client.get(wiki_settings.LOGOUT_URL)
+        user = auth.get_user(self.client)
+        # Can be changed to assertIs(... False) when Django 1.11 is dropped
+        # Django 1.11 returns CallableBoolean
+        self.assertFalse(auth.get_user(self.client).is_authenticated)
+        self.assertIsInstance(user, AnonymousUser)
+
+
+@wiki_override_settings(ACCOUNT_HANDLING=True)
+class LoginTestViews(RequireRootArticleMixin, TestBase):
+    def test_already_signed_in(self):
+        self.client.force_login(self.superuser1)
+        response = self.client.get(wiki_settings.LOGIN_URL)
+        self.assertRedirects(response, reverse('wiki:root'))
+
+    def test_log_in(self):
+        self.client.post(
+            wiki_settings.LOGIN_URL,
+            {'username': SUPERUSER1_USERNAME, 'password': SUPERUSER1_PASSWORD}
+        )
+        # Can be changed to assertIs(... False) when Django 1.11 is dropped
+        # Django 1.11 returns CallableBoolean
+        self.assertTrue(self.superuser1.is_authenticated)
+        self.assertEqual(auth.get_user(self.client), self.superuser1)
+
+
+class SignupViewTests(RequireRootArticleMixin, TestBase):
+
+    @wiki_override_settings(ACCOUNT_HANDLING=True, ACCOUNT_SIGNUP_ALLOWED=True)
+    def test_signup(self):
+        response = self.client.post(
+            wiki_settings.SIGNUP_URL,
+            data={
+                'password1': 'wiki', 'password2': 'wiki', 'username': 'wiki',
+                'email': 'wiki@wiki.com'
+            }
+        )
+        self.assertIs(CustomUser.objects.filter(email='wiki@wiki.com').exists(), True)
+        self.assertRedirects(response, reverse('wiki:login'))
